Fix volume value display
Since MPD does not send an update when a volume change happens, make the client store the new volume value in its state array every time a change is made.

This is a follow-up fix to 3f02cb0f12a16c091ca03cc2ebe2a4ad13151b44. Here's a bit of background to this problem. When Spop is playing and the user changes the volume, MPD detects the volume change and sends its full state array to Volumio, which includes specifying that the seek time is 0, since MPD is not playing. This causes Volumio to reset its seek bar to 0 as well, even though Spop is still playing at the original seek position.

In commit 3f02cb0f12a16c091ca03cc2ebe2a4ad13151b44, I've opted to not have Volumio respond to MPD updates when the only change is a volume change. This fixes the seek bar jumping when Spop is playing, but the downside is that other connected clients do not get notified when a volume change occurs, so their volume bars do not update.

A potential long term solution would be to rework the MPD state update code to just send the portion of the status which changed, and then have the Volumio UI only change the portions of its controls affected by that update.