Fix hang on backendRequest error
_player_engine.php blocks indefinately when the state sent by the client matches the current state.
But, in the case where the backendRequest hits an error, the app starts trying to reconnect in a
setTimeout loop while it shows a modal UI. So, we don't want _player_engine to block.

This adds an option to backendRequest to send state or not for the cases we want it to block or not.